Add and use AS chain index.
Index points into the list of ASs in the chain, and is used and
incremented appropriately.

Pretty-print AsChain objects nicely.

Remove a hacky way of determining the session case - we now carry
it in the AsChain and use that.

Nuance the no-fork handling a little. Still not perfect.

Support for deleting a header.