Split out AsChainStep from AsChain.
The AsChain is long-lived and immutable.

The AsChainStep is a short-lived immutable data class that points into it.

A key advantage of this model is that we have no index being incremented
internally - we simply tell each AS (by ODI token) the next index it
should ask Sprout to proceed to, and when we receive that request we
do it. That means we can cope with forking ASs, which is a win wrt
MSF-IA-SIP.017 which recommends we set Request-Disposition: no-fork
on all steps except the last.