Always examine tdata, not rdata.
Previously we sometimes examined rdata when it was available, e.g.,
when constructing an AsChain. With this fix, we no longer do that
- we always look at our current copy, tdata.

This is required in order to handle chains of internal ASs: an
internal AS may modify the message, and this must affect future
iFC evaluations etc - so we must always examine the latest copy,
which is in tdata.