Fix MMTEL orig-cdiv handling.
The internal MMTEL AS (in terminating mode) now holds an ODI token (an AsChainLink, actually). This enables it to correctly trigger
orig-cdiv handling when it does a redirect.

Fix the orig-cdiv handling to get the served user from the AsChain we have in our hands, rather than scraping the message for it -
previously this relied on the P-Served-User still being present in the message when the AS returns it, which is not guaranteed,
and now (in the internal case) the P-Served-User header might never have been added!