Translate XML entities in iFCs, and handle CDATA nodes properly.
We recently tried to just treat XML values as just text strings,
and not decode any &gt; or similar entities in them, on the basis
that we don't encode them when inserting iFCs. But this means we
have no way to express a literal "<", and so can't match SIP
headers with a value of e.g. "<sip:example.com>". This commit reverts us to the old position.