Add logs to the Sproutlet plugins.
Previously, there was no definite way to tell whether each plugin
thought it was enabled or not at load time. It was sometimes possible
to infer from the presence of other logs, but not always. This commit
adds a log to each to make this clear. It's fine to put this log in the
mainline as it only runs on startup, and is unlikely to get spammy.