Reduce the TCP timers to 1.8 seconds.
Currently the Session Continued AS timeout is identical to the TCP connection
timeout. This means that are much more like to see race conditions related to
the AS failing as the TCP timer and the AS timer fire at the same time.

This reduces the default TCP timer to 1.8 seconds.

We need > 1 s, as Linux will only send retransmit a SYN after 1 second, which
otherwise makes us vulnerable to dropped packets causing erroneous failure.

This means we should avoid the race condition one way or the other.